+++ src/main/java/org/matsim/freight/logistics/resourceImplementations/DistributionCarrierScheduler.java
@@ -1,4 +1,25 @@
 /*
+  *********************************************************************** *
+  * project: org.matsim.*
+  *                                                                         *
+  * *********************************************************************** *
+  *                                                                         *
+  * copyright       :  (C) 2024 by the members listed in the COPYING,       *
+  *                   LICENSE and WARRANTY file.                            *
+  * email           : info at matsim dot org                                *
+  *                                                                         *
+  * *********************************************************************** *
+  *                                                                         *
+  *   This program is free software; you can redistribute it and/or modify  *
+  *   it under the terms of the GNU General Public License as published by  *
+  *   the Free Software Foundation; either version 2 of the License, or     *
+  *   (at your option) any later version.                                   *
+  *   See also COPYING, LICENSE and WARRANTY file                           *
+  *                                                                         *
+  * ***********************************************************************
+ */
+
+/*
  *  *********************************************************************** *
  *  * project: org.matsim.*
  *  * *********************************************************************** *
@@ -18,7 +39,7 @@
  *  * ***********************************************************************
  */
 
-package org.matsim.freight.logistics.resourceImplementations.distributionCarrier;
+package org.matsim.freight.logistics.resourceImplementations;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -32,8 +53,6 @@ import org.matsim.freight.carriers.Tour.Leg;
 import org.matsim.freight.carriers.Tour.ServiceActivity;
 import org.matsim.freight.carriers.Tour.TourElement;
 import org.matsim.freight.logistics.*;
-import org.matsim.freight.logistics.resourceImplementations.CarrierSchedulerUtils;
-import org.matsim.freight.logistics.resourceImplementations.ResourceImplementationUtils;
 import org.matsim.freight.logistics.shipment.ShipmentPlanElement;
 import org.matsim.freight.logistics.shipment.ShipmentUtils;
 import org.matsim.vehicles.VehicleType;
