+++ src/main/java/org/matsim/freight/logistics/resourceImplementations/DistributionCarrierUtils.java
@@ -1,4 +1,25 @@
-package org.matsim.freight.logistics.resourceImplementations.distributionCarrier;
+/*
+  *********************************************************************** *
+  * project: org.matsim.*
+  *                                                                         *
+  * *********************************************************************** *
+  *                                                                         *
+  * copyright       :  (C) 2024 by the members listed in the COPYING,       *
+  *                   LICENSE and WARRANTY file.                            *
+  * email           : info at matsim dot org                                *
+  *                                                                         *
+  * *********************************************************************** *
+  *                                                                         *
+  *   This program is free software; you can redistribute it and/or modify  *
+  *   it under the terms of the GNU General Public License as published by  *
+  *   the Free Software Foundation; either version 2 of the License, or     *
+  *   (at your option) any later version.                                   *
+  *   See also COPYING, LICENSE and WARRANTY file                           *
+  *                                                                         *
+  * ***********************************************************************
+ */
+
+package org.matsim.freight.logistics.resourceImplementations;
 
 import java.util.ArrayList;
 import org.matsim.api.core.v01.Id;
@@ -7,7 +28,6 @@ import org.matsim.api.core.v01.network.Network;
 import org.matsim.freight.carriers.Carrier;
 import org.matsim.freight.logistics.LSPResource;
 import org.matsim.freight.logistics.LogisticChainElement;
-import org.matsim.freight.logistics.resourceImplementations.ResourceImplementationUtils;
 
 /**
  * @author Kai Martins-Turner (kturner)
